# Conflicts:
#	app/src/main/java/it/niedermann/owncloud/notes/android/appwidget/SingleNoteWidgetFactory.java
#	app/src/main/java/it/niedermann/owncloud/notes/android/fragment/NoteEditFragment.java
#	app/src/main/java/it/niedermann/owncloud/notes/android/fragment/NotePreviewFragment.java
#	app/src/main/java/it/niedermann/owncloud/notes/android/fragment/NoteReadonlyFragment.java
#	app/src/main/java/it/niedermann/owncloud/notes/util/MarkDownUtil.java
#	fastlane/metadata/android/en-US/images/icon.png

+++ app/src/main/java/it/niedermann/owncloud/notes/android/fragment/BaseNoteFragment.java
@@ -3,9 +3,9 @@ package it.niedermann.owncloud.notes.android.fragment;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.content.pm.ShortcutInfo;
 import android.content.pm.ShortcutManager;
+import android.graphics.Color;
 import android.graphics.drawable.Icon;
 import android.os.Build;
 import android.os.Bundle;
@@ -13,11 +13,14 @@ import android.util.Log;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
+import android.view.View;
+import android.widget.ScrollView;
 
+import androidx.annotation.ColorInt;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.appcompat.widget.ShareActionProvider;
-import androidx.core.view.MenuItemCompat;
+import androidx.core.content.ContextCompat;
+import androidx.fragment.app.DialogFragment;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 
@@ -29,18 +32,26 @@ import com.nextcloud.android.sso.model.SingleSignOnAccount;
 import it.niedermann.owncloud.notes.R;
 import it.niedermann.owncloud.notes.android.activity.EditNoteActivity;
 import it.niedermann.owncloud.notes.android.fragment.CategoryDialogFragment.CategoryDialogListener;
+import it.niedermann.owncloud.notes.android.fragment.EditTitleDialogFragment.EditTitleListener;
+import it.niedermann.owncloud.notes.branding.BrandedFragment;
+import it.niedermann.owncloud.notes.model.ApiVersion;
 import it.niedermann.owncloud.notes.model.CloudNote;
 import it.niedermann.owncloud.notes.model.DBNote;
 import it.niedermann.owncloud.notes.model.DBStatus;
 import it.niedermann.owncloud.notes.model.ISyncCallback;
 import it.niedermann.owncloud.notes.model.LocalAccount;
 import it.niedermann.owncloud.notes.persistence.NotesDatabase;
+import it.niedermann.owncloud.notes.util.ColorUtil;
 import it.niedermann.owncloud.notes.util.NoteUtil;
+import it.niedermann.owncloud.notes.util.ShareUtil;
 
 import static androidx.core.content.pm.ShortcutManagerCompat.isRequestPinShortcutSupported;
 import static it.niedermann.owncloud.notes.android.activity.EditNoteActivity.ACTION_SHORTCUT;
+import static it.niedermann.owncloud.notes.branding.BrandingUtil.tintMenuIcon;
+import static it.niedermann.owncloud.notes.util.ColorUtil.isColorDark;
+import static it.niedermann.owncloud.notes.util.Notes.isDarkThemeActive;
 
-public abstract class BaseNoteFragment extends Fragment implements CategoryDialogListener {
+public abstract class BaseNoteFragment extends BrandedFragment implements CategoryDialogListener, EditTitleListener {
 
     private static final String TAG = BaseNoteFragment.class.getSimpleName();
 
@@ -56,10 +67,13 @@ public abstract class BaseNoteFragment extends Fragment implements CategoryDialo
     private SingleSignOnAccount ssoAccount;
 
     protected DBNote note;
+    // TODO do we really need this? The reference to note is currently the same
     @Nullable
     private DBNote originalNote;
+    private int originalScrollY;
     protected NotesDatabase db;
     private NoteFragmentListener listener;
+    private boolean titleModified = false;
 
     protected boolean isNew = true;
 
@@ -88,7 +102,7 @@ public abstract class BaseNoteFragment extends Fragment implements CategoryDialo
                         if (content == null) {
                             throw new IllegalArgumentException(PARAM_NOTE_ID + " is not given, argument " + PARAM_NEWNOTE + " is missing and " + PARAM_CONTENT + " is missing.");
                         } else {
-                            note = new DBNote(-1, -1, null, NoteUtil.generateNoteTitle(content), content, false, getString(R.string.category_readonly), null, DBStatus.VOID, -1, "");
+                            note = new DBNote(-1, -1, null, NoteUtil.generateNoteTitle(content), content, false, getString(R.string.category_readonly), null, DBStatus.VOID, -1, "", 0);
                         }
                     } else {
                         note = db.getNote(localAccount.getId(), db.addNoteAndSync(ssoAccount, localAccount.getId(), cloudNote));
@@ -102,6 +116,32 @@ public abstract class BaseNoteFragment extends Fragment implements CategoryDialo
             setHasOptionsMenu(true);
         } catch (NextcloudFilesAppAccountNotFoundException | NoCurrentAccountSelectedException e) {
             e.printStackTrace();
+        }
+    }
+
+    @Nullable
+    protected abstract ScrollView getScrollView();
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+        final ScrollView scrollView = getScrollView();
+        if (scrollView != null) {
+            scrollView.getViewTreeObserver().addOnScrollChangedListener(() -> {
+                if (scrollView.getScrollY() > 0) {
+                    note.setScrollY(scrollView.getScrollY());
+                }
+            });
+        }
+    }
+
+    @Override
+    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+        final ScrollView scrollView = getScrollView();
+        if (scrollView != null) {
+            this.originalScrollY = note.getScrollY();
+            scrollView.post(() -> scrollView.scrollTo(0, originalScrollY));
         }
     }
 
@@ -143,12 +183,14 @@ public abstract class BaseNoteFragment extends Fragment implements CategoryDialo
     }
 
     @Override
-    public void onCreateOptionsMenu(@NonNull Menu menu, MenuInflater inflater) {
+    public void onCreateOptionsMenu(@NonNull Menu menu, @NonNull MenuInflater inflater) {
         inflater.inflate(R.menu.menu_note_fragment, menu);
 
         if (isRequestPinShortcutSupported(requireActivity()) && Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             menu.add(Menu.NONE, MENU_ID_PIN, 110, R.string.pin_to_homescreen);
         }
+
+        super.onCreateOptionsMenu(menu, inflater);
     }
 
     @Override
@@ -157,12 +199,14 @@ public abstract class BaseNoteFragment extends Fragment implements CategoryDialo
         MenuItem itemFavorite = menu.findItem(R.id.menu_favorite);
         prepareFavoriteOption(itemFavorite);
 
+        menu.findItem(R.id.menu_title).setVisible(localAccount.getPreferredApiVersion() != null && localAccount.getPreferredApiVersion().compareTo(new ApiVersion("1.0", 1, 0)) >= 0);
         menu.findItem(R.id.menu_delete).setVisible(!isNew);
     }
 
     private void prepareFavoriteOption(MenuItem item) {
         item.setIcon(note.isFavorite() ? R.drawable.ic_star_white_24dp : R.drawable.ic_star_border_white_24dp);
         item.setChecked(note.isFavorite());
+        tintMenuIcon(item, colorAccent);
     }
 
     /**
@@ -175,7 +219,7 @@ public abstract class BaseNoteFragment extends Fragment implements CategoryDialo
                 if (originalNote == null) {
                     db.deleteNoteAndSync(ssoAccount, note.getId());
                 } else {
-                    db.updateNoteAndSync(ssoAccount, localAccount.getId(), originalNote, null, null);
+                    db.updateNoteAndSync(ssoAccount, localAccount, originalNote, null, null);
                 }
                 listener.close();
                 return true;
@@ -191,24 +235,14 @@ public abstract class BaseNoteFragment extends Fragment implements CategoryDialo
             case R.id.menu_category:
                 showCategorySelector();
                 return true;
+            case R.id.menu_title:
+                showEditTitleDialog();
+                return true;
             case R.id.menu_move:
-                AccountChooserDialogFragment.newInstance().show(requireActivity().getSupportFragmentManager(), BaseNoteFragment.class.getCanonicalName());
+                MoveAccountDialogFragment.newInstance().show(requireActivity().getSupportFragmentManager(), BaseNoteFragment.class.getSimpleName());
                 return true;
             case R.id.menu_share:
-                Intent shareIntent = new Intent();
-                shareIntent.setAction(Intent.ACTION_SEND);
-                shareIntent.setType("text/plain");
-                shareIntent.putExtra(Intent.EXTRA_SUBJECT, note.getTitle());
-                shareIntent.putExtra(Intent.EXTRA_TEXT, note.getContent());
-
-
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-                    startActivity(Intent.createChooser(shareIntent, note.getTitle()));
-                } else {
-                    ShareActionProvider actionProvider = (ShareActionProvider) MenuItemCompat.getActionProvider(item);
-                    actionProvider.setShareIntent(shareIntent);
-                }
-
+                ShareUtil.openShareDialog(requireContext(), note.getTitle(), note.getContent());
                 return false;
             case MENU_ID_PIN:
                 if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
@@ -249,7 +283,7 @@ public abstract class BaseNoteFragment extends Fragment implements CategoryDialo
     }
 
     public void onCloseNote() {
-        if (originalNote == null && getContent().isEmpty()) {
+        if (!titleModified && originalNote == null && getContent().isEmpty()) {
             db.deleteNoteAndSync(ssoAccount, note.getId());
         }
     }
@@ -264,29 +298,19 @@ public abstract class BaseNoteFragment extends Fragment implements CategoryDialo
         if (note != null) {
             String newContent = getContent();
             if (note.getContent().equals(newContent)) {
-                Log.v(TAG, "... not saving, since nothing has changed");
+                if (note.getScrollY() != originalScrollY) {
+                    Log.v(TAG, "... only saving new scroll state, since content did not change");
+                    db.updateScrollY(note.getId(), note.getScrollY());
+                } else {
+                    Log.v(TAG, "... not saving, since nothing has changed");
+                }
             } else {
-                note = db.updateNoteAndSync(ssoAccount, localAccount.getId(), note, newContent, callback);
+                note = db.updateNoteAndSync(ssoAccount, localAccount, note, newContent, callback);
                 listener.onNoteUpdated(note);
+                requireActivity().invalidateOptionsMenu();
             }
         } else {
             Log.e(TAG, "note is null");
-        }
-    }
-
-    @SuppressWarnings("WeakerAccess") //PMD...
-    protected float getFontSizeFromPreferences(SharedPreferences sp) {
-        final String prefValueSmall = getString(R.string.pref_value_font_size_small);
-        final String prefValueMedium = getString(R.string.pref_value_font_size_medium);
-        // final String prefValueLarge = getString(R.string.pref_value_font_size_large);
-        String fontSize = sp.getString(getString(R.string.pref_key_font_size), prefValueMedium);
-
-        if (fontSize.equals(prefValueSmall)) {
-            return getResources().getDimension(R.dimen.note_font_size_small);
-        } else if (fontSize.equals(prefValueMedium)) {
-            return getResources().getDimension(R.dimen.note_font_size_medium);
-        } else {
-            return getResources().getDimension(R.dimen.note_font_size_large);
         }
     }
 
@@ -311,15 +335,71 @@ public abstract class BaseNoteFragment extends Fragment implements CategoryDialo
         categoryFragment.show(manager, fragmentId);
     }
 
+    /**
+     * Opens a dialog in order to chose a category
+     */
+    private void showEditTitleDialog() {
+        final String fragmentId = "fragment_edit_title";
+        FragmentManager manager = requireActivity().getSupportFragmentManager();
+        Fragment frag = manager.findFragmentByTag(fragmentId);
+        if (frag != null) {
+            manager.beginTransaction().remove(frag).commit();
+        }
+        DialogFragment editTitleFragment = EditTitleDialogFragment.newInstance(note.getTitle());
+        editTitleFragment.setTargetFragment(this, 0);
+        editTitleFragment.show(manager, fragmentId);
+    }
+
     @Override
     public void onCategoryChosen(String category) {
         db.setCategory(ssoAccount, note, category, null);
         listener.onNoteUpdated(note);
     }
 
+    @Override
+    public void onTitleEdited(String newTitle) {
+        titleModified = true;
+        note.setTitle(newTitle);
+        note = db.updateNoteAndSync(ssoAccount, localAccount, note, note.getContent(), newTitle, null);
+        listener.onNoteUpdated(note);
+    }
+
     public void moveNote(LocalAccount account) {
         db.moveNoteToAnotherAccount(ssoAccount, note.getAccountId(), note, account.getId());
         listener.close();
+    }
+
+    @ColorInt
+    protected static int getTextHighlightBackgroundColor(@NonNull Context context, @ColorInt int mainColor, @ColorInt int colorPrimary, @ColorInt int colorAccent) {
+        if (isDarkThemeActive(context)) { // Dark background
+            if (isColorDark(mainColor)) { // Dark brand color
+                if (ColorUtil.contrastRatioIsSufficient(mainColor, colorPrimary)) { // But also dark text
+                    return mainColor;
+                } else {
+                    return ContextCompat.getColor(context, R.color.defaultTextHighlightBackground);
+                }
+            } else { // Light brand color
+                if (ColorUtil.contrastRatioIsSufficient(mainColor, colorAccent)) { // But also dark text
+                    return Color.argb(77, Color.red(mainColor), Color.green(mainColor), Color.blue(mainColor));
+                } else {
+                    return ContextCompat.getColor(context, R.color.defaultTextHighlightBackground);
+                }
+            }
+        } else { // Light background
+            if (isColorDark(mainColor)) { // Dark brand color
+                if (ColorUtil.contrastRatioIsSufficient(mainColor, colorAccent)) { // But also dark text
+                    return Color.argb(77, Color.red(mainColor), Color.green(mainColor), Color.blue(mainColor));
+                } else {
+                    return ContextCompat.getColor(context, R.color.defaultTextHighlightBackground);
+                }
+            } else { // Light brand color
+                if (ColorUtil.contrastRatioIsSufficient(mainColor, colorPrimary)) { // But also dark text
+                    return mainColor;
+                } else {
+                    return ContextCompat.getColor(context, R.color.defaultTextHighlightBackground);
+                }
+            }
+        }
     }
 
     public interface NoteFragmentListener {
